introduced weighted.fitting variable
This variable controls the type of non linear fitting that is performed.
If TRUE then weights equal to the inverse of the variances of the values
being fitted are used, if FALSE ( default ) then weightless fitting is
performed instead.

Additionally, the message reporting unconverged data has been altered so
that it prints the maximum of the max.of.mins instead of the a0 of the
fitting of the max.of.mins values.