Fixed a bug in the check for binary files function.
Older versions of R don't include the 'raw=TRUE/FALSE' flag for the
'file' function and since it hasn't come up as necessary in any of the
tests i run, i have removed it so that the script can run without
problems in machines running older R versions. The bug was identified in
a machine running R 2.10.1.